// src/app/model/CrmDashboardAnalytics.ts
export class CrmDashboardAnalytics {
  newLeads: number = 0;
  recentlyAddedNewLeadsAmount: number = 0;
  closedLeads: number = 0;
  recentlyClosedLeadsAmount: number = 0;
  convertedLeads: number = 0;
  recentlyConvertedLeadsAmount: number = 0;
  possibleLeads: number = 0;
  recentlyPosibleLeadsAmount: number = 0;
  totalLeads: number = 0;
  cnpLeads: number = 0;
  leadStats: number[] = [];
}